cloud_functions/cp_ratio_alert/requirements.txt
google-cloud-storage>=2.0.0
google-cloud-pubsub>=2.0.0
google-cloud-secret-manager>=2.0.0
pandas>=1.3.0
numpy>=1.20.0